Add a shadow between the calendar and the agenda
