+++ agendacalendarview/src/main/java/com/github/tibolte/agendacalendarview/agenda/AgendaListView.java
@@ -41,67 +41,6 @@ public class AgendaListView extends StickyListHeadersListView {
 
     // endregion
 
-    // region Class - View
-
-    @Override
-    protected void onFinishInflate() {
-        super.onFinishInflate();
-
-        BusProvider.getInstance().toObserverable()
-                .subscribe(event -> {
-                    if (event instanceof Events.DayClickedEvent) {
-                        Events.DayClickedEvent clickedEvent = (Events.DayClickedEvent) event;
-                        scrollToCurrentDate(clickedEvent.getCalendar());
-                    } else if (event instanceof Events.CalendarScrolledEvent) {
-                        int offset = (int) (3 * getResources().getDimension(R.dimen.day_cell_height));
-                        translateList(offset);
-                    } else if (event instanceof Events.EventsFetched) {
-                        ((AgendaAdapter) getAdapter()).updateEvents(CalendarManager.getInstance().getEvents());
-
-                        getViewTreeObserver().addOnGlobalLayoutListener(
-                                new ViewTreeObserver.OnGlobalLayoutListener() {
-                                    @Override
-                                    public void onGlobalLayout() {
-                                        if (getWidth() != 0 && getHeight() != 0) {
-                                            // display only two visible rows on the calendar view
-                                            ViewGroup.MarginLayoutParams layoutParams = (ViewGroup.MarginLayoutParams) getLayoutParams();
-                                            int height = getHeight();
-                                            int margin = (int) (getContext().getResources().getDimension(R.dimen.calendar_header_height) + 2 * getContext().getResources().getDimension(R.dimen.day_cell_height));
-                                            layoutParams.height = height - margin;
-                                            layoutParams.setMargins(0, margin, 0, 0);
-                                            setLayoutParams(layoutParams);
-
-                                            scrollToCurrentDate(CalendarManager.getInstance().getToday());
-
-                                            getViewTreeObserver().removeOnGlobalLayoutListener(this);
-                                        }
-                                    }
-                                }
-                        );
-                    } else if (event instanceof Events.ForecastFetched) {
-                        ((AgendaAdapter) getAdapter()).updateEvents(CalendarManager.getInstance().getEvents());
-                    }
-                });
-    }
-
-    @Override
-    public boolean dispatchTouchEvent(MotionEvent event) {
-        int eventaction = event.getAction();
-
-        switch (eventaction) {
-            case MotionEvent.ACTION_DOWN:
-                // if the user touches the listView, we put it back to the top
-                translateList(0);
-                break;
-            default:
-                break;
-        }
-
-        return super.dispatchTouchEvent(event);
-    }
-
-    // endregion
-
     // region Public methods
 
     public void scrollToCurrentDate(Calendar today) {
@@ -117,35 +56,6 @@ public class AgendaListView extends StickyListHeadersListView {
 
         final int finalToIndex = toIndex;
         post(()->setSelection(finalToIndex));
-    }
-
-    public void translateList(int targetY) {
-        ObjectAnimator mover = ObjectAnimator.ofFloat(this, "translationY", targetY);
-        mover.setDuration(150);
-        if (targetY == 0) {
-            mover.addListener(new Animator.AnimatorListener() {
-                @Override
-                public void onAnimationStart(Animator animation) {
-
-                }
-
-                @Override
-                public void onAnimationEnd(Animator animation) {
-                    BusProvider.getInstance().send(new Events.AgendaListViewTouchedEvent());
-                }
-
-                @Override
-                public void onAnimationCancel(Animator animation) {
-
-                }
-
-                @Override
-                public void onAnimationRepeat(Animator animation) {
-
-                }
-            });
-        }
-        mover.start();
     }
 
     // endregion
