+++ agendacalendarview/src/main/res/drawable/shadow.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:angle="270"
+        android:startColor="#9C9CA0"
+        android:endColor="#F3F3F3" />
+</shape>
