+++ agendacalendarview/src/main/res/layout/view_agenda.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<merge
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <View
+        android:id="@+id/view_shadow"
+        android:layout_width="match_parent"
+        android:layout_height="5dp"
+        android:background="@drawable/shadow">
+    </View>
+
+    <com.github.tibolte.agendacalendarview.agenda.AgendaListView
+        android:id="@+id/agenda_listview"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:animationCache="false"
+        android:divider="@color/calendar_divider_color"
+        android:dividerHeight="1dp"
+        android:overScrollMode="never"
+        android:scrollbars="none"
+        android:scrollingCache="false"/>
+
+</merge>
